agent/requirements.txt
pandas>=2.3.0
scikit-learn>=1.7.0
joblib>=1.5.0
requests>=2.32.0
psycopg2-binary>=2.9.10
python-dotenv>=1.2.2